changed data type related to demands from integer to unsigned long long

+++ src/basic_functions.h
@@ -4,7 +4,7 @@ void allocate_results (struct result *R, struct demand *D);
 void deallocate_results	(struct result *R);
 
 
-void randomize_vector (int *vector);
+void randomize_vector (unsigned long long *vector);
 
 int count_flight_list (char *filename);
 void read_flight_list (char *filename, struct flight* F);
@@ -14,4 +14,4 @@ void deallocate_memory_network (struct network* G);
 
 void read_demand_matrix (char *filename, struct demand* D);
 void deallocate_memory_demand ( struct demand* D);
-void remove_pair_from_demand(int q, struct demand* D);
+void remove_pair_from_demand(unsigned long long q, struct demand* D);
